requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymysql==1.1.0
sqlalchemy==2.0.30
pymongo==4.7.2
python-dotenv==1.0.1
cryptography==42.0.7